Switch to a non-MetalLB setup
We can't easily expose LoadBalancer services on macOS because Docker
runs in a VM and the service's address isn't reachable from the host
which in turns makes verifiers fail.

Routing traffic to ingresses and Gateways isn't an "interesting" part in
the context of ingress2gateway: We care about the behavior of the GWAPI
resources once traffic gets to them, but we don't care about *how*
traffic gets to them. Therefore, exposing services via a load balancer
isn't a requirement for the tests.

Signed-off-by: Johanan Liebermann <jliebermann@microsoft.com>